test: Move API test to separate file
